+++ series/telegram_bot/06_SistemaVigilancia/arduino/SistemaVigilancia/token_ejemplo.h
@@ -1,5 +1,5 @@
 const char* ssid = "xxx";
 const char* password = "xxx";
 const String token = "xxx"; 
-const int IDchat = 1234;
+const int64_t IDchat = 1234;
 const String nombre = "Estudio";
